+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- `delete` command for deleting modules from Puppetfile.
+
 ## [0.4.0] - 2017-10-07
 ### Added
 - Support for setting `changeset` for git/hg modules.